feat(guardrails): ruff debt 541->0, pre-commit gates, crash-proof audit, M-61/L-65
- Ruff debt (plan 20260821): F821/F811 latent-NameError fixes, PLW1510 explicit check=False, per-site F841/SIM/TRY004/C40x/PYI034/RUF012/S112, 243 auto-fixes. B008 -> per-file-ignore (server routes/deps/tests); BLE001(114)+SIM117 extend-ignored as tracked debt with in-config rationale.
- Pre-commit gate: ruff + black(25.1.0, line-length 120) + mypy(2.3.1) + detect-secrets(baseline 403), alongside pre-existing bandit/uv-lock/pip-audit.
- Area-1 audit crash-proof: executor AuditLogger rewritten to durable disk-first spool (synchronous emit + daemon forwarder thread; serialization errors -> marker line, never dropped). shutdown() intentionally not wired (durability guaranteed by emit; no session-end hook).
- M-61 (FIXED): per-executor registration rate limit now enforced on its own persistent limiter (AND with per-IP, deny if either trips; headers reflect tighter). +2 tests + _LIMITERS isolation fixture.
- L-65 (FIXED): sbx chmod 400 fail-closed (check rc, roll back copied files, raise; matches sibling cp). +1 test.
- Tests: server 628->630, executor 528->534, core 313+7; ruff(black@25.1.0)/mypy gates green.
- Noted M-62/M-63 (two per-request-instantiated no-op registration limiters) as open in loose_ends_20260821.

+++ packages/core/src/core/engine/factory.py
@@ -1,11 +1,8 @@
 """CoreFactory: builder pattern for creating Core instances."""
 
-
-from pathlib import Path
-
 from .backend import Backend, BackendConfig
-from .rate_limiter import RateLimiter
 from .core import Core
+from .rate_limiter import RateLimiter
 
 
 class CoreFactoryError(Exception):
@@ -51,9 +48,7 @@ class CoreFactory:
         """
         database_url = config_dict.get("database_url")
         if database_url is None:
-            raise CoreFactoryError(
-                "config_dict missing required key: 'database_url'"
-            )
+            raise CoreFactoryError("config_dict missing required key: 'database_url'")
         passphrase = config_dict.get("passphrase")
         if isinstance(passphrase, str):
             passphrase = passphrase.encode("utf-8")
@@ -65,9 +60,7 @@ class CoreFactory:
         )
         return cls(config)
 
-    def with_rate_limiter(
-        self, max_attempts: int = 5, window_seconds: float = 300.0
-    ) -> CoreFactory:
+    def with_rate_limiter(self, max_attempts: int = 5, window_seconds: int = 300) -> CoreFactory:
         """Set rate limiter parameters.
 
         Args:
@@ -113,8 +106,7 @@ class CoreFactory:
                 self._kek = backend.bootstrap_kek()
             else:
                 raise CoreFactoryError(
-                    "KEK not available: provide either a passphrase or call "
-                    "with_kek() before build()"
+                    "KEK not available: provide either a passphrase or call " "with_kek() before build()"
                 )
 
         rate_limiter = self._rate_limiter or RateLimiter()
